pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "streakiller"
version = "0.1.0"
description = "Satellite streak detection pipeline for FITS astronomy images"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "opencv-python",
    "astropy",
    "skyfield",
    "requests",
    "matplotlib",
    "pillow",
    "click",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
]

[project.scripts]
streakiller = "streakiller.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.coverage.run]
source = ["src/streakiller"]
omit = ["tests/*", "legacy/*"]
